+++ packages/react-pose/src/components/Transition/children.ts
@@ -1,21 +1,12 @@
 import * as React from 'react';
 import { ReactElement } from 'react';
-import { CurrentPose } from '../components/PoseElement.types';
+import { CurrentPose } from '../PoseElement/types';
+import { MergeChildrenProps, Props } from './types';
 const { Children, cloneElement } = React;
 
-export interface MergeChildrenProps {
-  incomingChildren: Array<React.ReactElement<any>>;
-  displayedChildren: Array<React.ReactElement<any>>;
-  isLeaving: Set<string>;
-  removeFromTree: (key: string) => void;
-  preEnterPose: string;
-  enterPose: string;
-  exitPose: string;
-  flipMove: boolean;
-  animateOnMount: boolean;
-}
-
 const getKey = (child: ReactElement<any>): string => child.key as string;
+
+const filterChildProps = ({ children, _pose, onPoseComplete, popFromFlow, ...props }: Props) => props;
 
 const animateChildrenList = (
   incomingChildren: Array<ReactElement<any>>,
@@ -35,11 +26,17 @@ const mergeChildren = ({
   displayedChildren,
   isLeaving,
   removeFromTree,
-  preEnterPose,
-  enterPose,
-  exitPose,
-  flipMove
+  groupProps
 }: MergeChildrenProps) => {
+  const {
+    children: groupChildren,
+    preEnterPose,
+    enterPose,
+    exitPose,
+    flipMove,
+    animateOnMount,
+    ...propsForChild
+  } = groupProps;
   const children: Array<ReactElement<any>> = [];
 
   const prevKeys = displayedChildren.map(getKey);
@@ -65,11 +62,19 @@ const mergeChildren = ({
   );
 
   incomingChildren.forEach(child => {
-    const newChildProps = entering.has(child.key as string)
-      ? { initialPose: preEnterPose, _pose: enterPose }
-      : moving.has(child.key as string) && flipMove
-        ? { _pose: [enterPose, 'flip'] }
-        : { _pose: enterPose };
+    const newChildProps = {
+      ...propsForChild,
+      ...filterChildProps(child.props)
+    };
+
+    if (entering.has(child.key as string)) {
+      newChildProps.initialPose = preEnterPose;
+      newChildProps._pose = enterPose;
+    } else if (moving.has(child.key as string) && flipMove) {
+      newChildProps._pose = [enterPose, 'flip'];
+    } else {
+      newChildProps._pose = enterPose;
+    }
 
     children.push(cloneElement(child, newChildProps));
   });
@@ -79,17 +84,19 @@ const mergeChildren = ({
     const newChild = cloneElement(child, {
       _pose: exitPose,
       onPoseComplete: removeFromTree(key),
-      popFromFlow: flipMove
+      popFromFlow: flipMove,
+      ...propsForChild,
+      ...filterChildProps(child.props)
     });
 
-    let insertionIndex = prevKeys.indexOf(key);
+    const insertionIndex = prevKeys.indexOf(key);
 
     // We might have had new items added before this item in the same
     // render. So here we find the correct item to anchor to. This is
     // a pretty shitty algo. But it is also the one we have
-    if (insertionIndex) {
-      // TODO: Write a shitty algo
-    }
+    // if (insertionIndex) {
+    // TODO: Write a shitty algo
+    // }
 
     children.splice(insertionIndex, 0, newChild);
   });
@@ -98,13 +105,8 @@ const mergeChildren = ({
 };
 
 export const handleIncomingChildren = (props: MergeChildrenProps) => {
-  const {
-    displayedChildren,
-    incomingChildren,
-    animateOnMount,
-    preEnterPose,
-    enterPose
-  } = props;
+  const { displayedChildren, incomingChildren, groupProps } = props;
+  const { animateOnMount, preEnterPose, enterPose } = groupProps;
 
   // If initial mount and we're animating
   if (!displayedChildren && animateOnMount) {