+++ packages/react-pose/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 React Pose adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.3.3] 2018-09-17
+
+### Fixed
+
+- Filtering props from children that may have previously been set by `PoseGroup` itself.
+
+## [3.3.2] 2018-09-17
+
+### Fixed
+
+- Overriding `Transition` props with child props.
+
+## [3.3.1] 2018-09-17
+
+### Added
+
+- `Transition` now forwards children to props.
+
 ## [3.3.0] 2018-09-06
 
 ### Added